[AIP-21] Move operators.druid_check_operator providers.apache.druid.operators.druid_check
