+++ scripts/ci/pylint_todo.txt
@@ -142,7 +142,7 @@
 ./airflow/operators/bash_operator.py
 ./airflow/operators/check_operator.py
 ./airflow/operators/dagrun_operator.py
-./airflow/operators/druid_check_operator.py
+./airflow/providers/apache/druid/operators/druid_check.py
 ./airflow/operators/dummy_operator.py
 ./airflow/operators/email_operator.py
 ./airflow/operators/generic_transfer.py
